initial change to generic keys for hash table

+++ engine/src/core/dm_mem.c
@@ -5,30 +5,30 @@
 #include <string.h>
 #include <stdio.h>
 
- typedef struct dm_mem_db
+typedef struct dm_mem_db
 {
 	size_t allocs[DM_MEM_UNKNOWN];
 	size_t total;
 } dm_mem_db;
 
- static const char* mem_tag_str[] = {
-	 "ENGINE         ",
-	 "PLATFORM       ",
-	 "APPLICATION    ",
-	 "INPUT          ",
-	 "STRING         ",
-	 "LIST           ",
-	 "MAP            ",
-	 "RENDERER       ",
-	 "RENDER COMMAND ",
-	 "RENDER PIPELINE",
-	 "RENDER PASS    ",
-	 "BUFFER         ",
-	 "SHADER         ",
-	 "UNIFORM        ",
-	 "TEXTURE        ",
-	 "UNKNOWN        "
- };
+static const char* mem_tag_str[] = {
+    "ENGINE         ",
+    "PLATFORM       ",
+    "APPLICATION    ",
+    "INPUT          ",
+    "STRING         ",
+    "LIST           ",
+    "MAP            ",
+    "RENDERER       ",
+    "RENDER COMMAND ",
+    "RENDER PIPELINE",
+    "RENDER PASS    ",
+    "BUFFER         ",
+    "SHADER         ",
+    "UNIFORM        ",
+    "TEXTURE        ",
+    "UNKNOWN        "
+};
 
 dm_mem_db mem_db = { 0 };
 
@@ -36,7 +36,7 @@ void* dm_alloc(size_t size, dm_mem_tag tag)
 {
 	mem_db.total += size;
 	mem_db.allocs[tag] += size;
-
+    
 	return dm_platform_alloc(size);
 }
 
@@ -44,7 +44,7 @@ void* dm_calloc(size_t count, size_t size, dm_mem_tag tag)
 {
 	mem_db.total += size * count;
 	mem_db.allocs[tag] += size * count;
-
+    
 	return dm_platform_calloc(count, size);
 }
 
@@ -57,7 +57,7 @@ void dm_free(void* block, size_t size, dm_mem_tag tag)
 {
 	mem_db.total -= size;
 	mem_db.allocs[tag] -= size;
-
+    
 	dm_platform_free(block);
 }
 
@@ -80,23 +80,23 @@ void dm_mem_db_adjust(size_t size, dm_mem_tag tag, dm_mem_adjust_func adjust_fun
 {
 	switch (adjust_func)
 	{
-	case DM_MEM_ADJUST_ADD:
+        case DM_MEM_ADJUST_ADD:
 		mem_db.total += size;
 		mem_db.allocs[tag] += size;
 		break;
-	case DM_MEM_ADJUST_SUBTRACT:
+        case DM_MEM_ADJUST_SUBTRACT:
 		mem_db.total -= size;
 		mem_db.allocs[tag] -= size;
 		break;
-	case DM_MEM_ADJUST_MULTIPLY:
+        case DM_MEM_ADJUST_MULTIPLY:
 		mem_db.total += (mem_db.allocs[tag] * size);
 		mem_db.allocs[tag] *= size;
 		break;
-	case DM_MEM_ADJUST_DIVIDE:
+        case DM_MEM_ADJUST_DIVIDE:
 		mem_db.total += (mem_db.allocs[tag] / size);
 		mem_db.allocs[tag] *= size;
 		break;
-	default: break;
+        default: break;
 	}
 	
 }
@@ -106,41 +106,41 @@ void dm_mem_track()
 	const size_t gb = 1024 * 1024 * 1024;
 	const size_t mb = 1024 * 1024;
 	const size_t kb = 1024;
-
+    
 	char buffer[1024] = "System memory usage: \n";
 	size_t offset = strlen(buffer);
-
+    
 	for (int i = 0; i < DM_MEM_UNKNOWN; i++)
 	{
 		char unit[4] = "XiB";
 		float amount = 1.0f;
-
+        
 		if (mem_db.allocs[i] >= gb)
 		{
 			unit[0] = 'G';
-			amount = mem_db.allocs[i] / (float)gb;
+			amount = (float)mem_db.allocs[i] / (float)gb;
 		}
 		else if (mem_db.allocs[i] >= mb)
 		{
 			unit[0] = 'M';
-			amount = mem_db.allocs[i] / (float)mb;
+			amount = (float)mem_db.allocs[i] / (float)mb;
 		}
 		else if (mem_db.allocs[i] >= kb)
 		{
 			unit[0] = 'K';
-			amount = mem_db.allocs[i] / (float)kb;
+			amount = (float)mem_db.allocs[i] / (float)kb;
 		}
 		else 
 		{
 			unit[0] = 'B';
 			unit[1] = 0;
-			amount = mem_db.allocs[i];
+			amount = (float)mem_db.allocs[i];
 		}
-
+        
 		int len = snprintf(buffer + offset, 1024, "  %s: %.2f%s\n", mem_tag_str[i], amount, unit);
 		offset += len;
 	}
-
+    
 	DM_LOG_INFO("%s", buffer);
 }
 
@@ -150,58 +150,58 @@ void dm_mem_all_freed()
 	{
 		DM_LOG_WARN("You have a memory leak somewhere! Total allocated memory is not 0!");
 		DM_LOG_WARN("You should... ");
-
+        
 		for (int i = 0; i < DM_MEM_UNKNOWN; i++)
 		{
 			char buffer[512];
-
+            
 			if (mem_db.allocs[i] > 0)
 			{
 				switch (i)
 				{
-				case DM_MEM_ENGINE:
+                    case DM_MEM_ENGINE:
 					strcpy(buffer, "engine");
 					break;
-				case DM_MEM_PLATFORM:
+                    case DM_MEM_PLATFORM:
 					strcpy(buffer, "platform");
 					break;
-				case DM_MEM_STRING:
+                    case DM_MEM_STRING:
 					strcpy(buffer, "string");
 					break;
-				case DM_MEM_LIST:
+                    case DM_MEM_LIST:
 					strcpy(buffer, "list");
 					break;
-				case DM_MEM_MAP:
+                    case DM_MEM_MAP:
 					strcpy(buffer, "map");
 					break;
-				case DM_MEM_RENDERER:
+                    case DM_MEM_RENDERER:
 					strcpy(buffer, "renderer");
 					break;
-				case DM_MEM_RENDER_COMMAND:
+                    case DM_MEM_RENDER_COMMAND:
 					strcpy(buffer, "render command");
 					break;
-				case DM_MEM_RENDER_PIPELINE:
+                    case DM_MEM_RENDER_PIPELINE:
 					strcpy(buffer, "render pipeline");
 					break;
-				case DM_MEM_RENDER_PASS:
+                    case DM_MEM_RENDER_PASS:
 					strcpy(buffer, "render pass");
 					break;
-				case DM_MEM_RENDERER_BUFFER:
+                    case DM_MEM_RENDERER_BUFFER:
 					strcpy(buffer, "buffer");
 					break;
-				case DM_MEM_RENDERER_SHADER:
+                    case DM_MEM_RENDERER_SHADER:
 					strcpy(buffer, "shader");
 					break;
-				case DM_MEM_RENDERER_UNIFORM:
+                    case DM_MEM_RENDERER_UNIFORM:
 					strcpy(buffer, "uniform");
 					break;
-				case DM_MEM_RENDERER_TEXTURE:
+                    case DM_MEM_RENDERER_TEXTURE:
 					strcpy(buffer, "texture");
 					break;
-				default:
+                    default:
 					break;
 				}
-
+                
 				DM_LOG_WARN("Check %s allocations...", buffer);
 			}
 		}
